backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
Deprecated==1.2.15
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
fonttools==4.55.0
idna==3.10
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.9.3
multipledispatch==1.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pykrx==1.0.48
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.17.0
xlrd==2.0.1
zope.interface==7.2
